feat: Implement new UI layouts, data grid columns, and export functionalities for various logistics processes and reports.

+++ src/contextLogistics/processes/warehouseEntry/ui/GridCol/columns.tsx
@@ -1,37 +1,63 @@
-import { GridActionsCellItem, GridColDef } from '@mui/x-data-grid'
-import { getDataState, WarehouseEntryENTITY } from 'logiflowerp-sdk'
+import { GridActionsCellItem, GridActionsCellItemProps, GridColDef } from '@mui/x-data-grid'
+import { getDataState, StateOrder, WarehouseEntryENTITY } from 'logiflowerp-sdk'
 import { CustomStatusOrder } from '@shared/ui-library'
 import DeleteForeverRoundedIcon from '@mui/icons-material/DeleteForeverRounded';
 import EditIcon from '@mui/icons-material/Edit'
+import { ReactElement } from 'react'
+import PictureAsPdfIcon from '@mui/icons-material/PictureAsPdf'
 
 interface IParams {
     handleDeleteClick: (row: WarehouseEntryENTITY) => void
     handleEditClick: (row: WarehouseEntryENTITY) => void
-    canDeleteWarehouseEntryByID: boolean
+    canDeleteWarehouseEntryByID: boolean,
+    handleViewPdfClick: (row: WarehouseEntryENTITY) => void
 }
 
 export const columns = (params: IParams): GridColDef<WarehouseEntryENTITY>[] => {
-    const { handleEditClick, handleDeleteClick, canDeleteWarehouseEntryByID } = params
+    const {
+        handleEditClick,
+        handleDeleteClick,
+        canDeleteWarehouseEntryByID,
+        handleViewPdfClick
+    } = params
     return [
         {
             field: 'Acciones',
             type: 'actions',
-            getActions: (params) => [
-                <GridActionsCellItem
-                    icon={<EditIcon color='info' />}
-                    label='Editar'
-                    onClick={() => handleEditClick(params.row)}
-                    showInMenu
-                />,
-                canDeleteWarehouseEntryByID
-                    ? <GridActionsCellItem
-                        icon={<DeleteForeverRoundedIcon color='error' />}
-                        label='Eliminar'
-                        onClick={() => handleDeleteClick(params.row)}
-                        showInMenu
-                    />
-                    : <></>,
-            ],
+            getActions: (params) => {
+                const actions: ReactElement<GridActionsCellItemProps>[] = []
+                if (params.row.state !== StateOrder.VALIDADO) {
+                    actions.push(
+                        <GridActionsCellItem
+                            icon={<EditIcon color='info' />}
+                            label='Editar'
+                            onClick={() => handleEditClick(params.row)}
+                            showInMenu
+                        />
+                    )
+                    if (canDeleteWarehouseEntryByID) {
+                        actions.push(
+                            <GridActionsCellItem
+                                icon={<DeleteForeverRoundedIcon color='error' />}
+                                label='Eliminar'
+                                onClick={() => handleDeleteClick(params.row)}
+                                showInMenu
+                            />
+                        )
+                    }
+                }
+                if (params.row.state === StateOrder.VALIDADO) {
+                    actions.push(
+                        <GridActionsCellItem
+                            icon={<PictureAsPdfIcon color='info' />}
+                            label='Ver guía PDF'
+                            onClick={() => handleViewPdfClick(params.row)}
+                            showInMenu
+                        />
+                    )
+                }
+                return actions
+            },
         },
         {
             field: 'documentNumber',
